Skip smallest primes during sieving
It is known that smallest primes contribute a significant
part of the cost of sieving. We ignore them while accumulating
the sum of log(p), and restore the missing part when scanning
for smooth candidates.

For a 220-bit semiprime, the cost of SIQS is reduced from
294G to 256G instructions, and MPQS goes from 446G instructions
to 399G. This translates to an approximate 10-15% speed gain.