Use memchr library to scan very large primes.
When handling very large primes (over 4x block size) it is
unlikely that a given prime number participates in the sieve
result. This means we have to scan for sparse zero bytes inside
a large array. The memchr crate provides SIMD-powered byte
lookups.

On a 240 bit input, the total performance improvement
can reach 25%.