Rework logarithms during sieve
The base 2 logarithm of the factor base takes only discrete
values between 2 and 20-24. It is not very efficient to keep
all the log2 values for every element of the factor base.

Since it is piecewise constant, we only need to know the intervals
over which log2 takes a given value. During the hot parts of the
sieve code, this saves a load from memory because the log can be
kept in a register or on stack.

On very large inputs (over 240 bits), this can provide a speedup
of 10-15% by reducing instruction count and increasing IPC.