Overhaul how large primes are sieved
The previous implementation was iterating multiple
times through the same array of large primes for each block,
but when primes are really large, they may hit several blocks
later or even never in the given interval.

We can instead fill the "hash table" in advance for the whole
interval and use the compact information during the sieve
without duplicate work.

This costs typically more instructions but is more memory I/O
friendly resulting in higher IPC on a modern CPU.

The gains are variable but can reach 5-20% of program running time.