Much faster initialization of SIQS polynomials.
Instead of reducing each 2^k values of coefficient B
modulo the factor base, precompute reductions mod p
for the k basic terms from which the 2^k values are generated.

The usual way is to use a Gray code to perform a single
addition per polynomial switch. However in practice
the cost of summing large vectors is very small in the total cost.

This can make polynomial preparation up to 2x faster and
makes the entire program 10-20% faster especially for
large inputs.