Make large factor bases more memory I/O friendly
For large inputs (320 bits) the sieve interval size can be 1 million
and the factor base size can be above 160k elements.

In this situation the size of the "hash table" of large prime sieve hits
is 8 bytes for sieve interval elements, which is too much for the L2/L3
cache of various CPU models.

By compacting necessary information as a bit field, each entry can be
reduced to 32 bits, halving the memory footprint.

For a 320-bit input, this results in a typical instruction count increase by 10%
compensated by a 30%-40% increase in instructions per cycle.