Implement "double large prime" strategy for MPQS and SIQS
This patch refactors how relations are handled by defining
a new container type with high-level methods (reducing
repeated boilerplate).

In the MPQS and SIQS/HMPQS methods, a new threshold
(256-bit input) triggers the double large prime strategy,
with a lower large prime threshold but twice larger
allowance for the size of smooth numbers.

The other parameters (size of factor base, size of intervals)
are not/almost not modified.