More aggressive precomputations in SIQS
The bulk of polynomial preparation step is to compute
(r-B)/A mod p for all primes of the factor base.

Instead of precomputing r, Bi, 1/A for a given A,
we can precompute r/A and Bi/A avoiding a couple
of extra calculations for each polynomial.

This saves about 40% of the preparation time, the
overall speed improvement can reach about 5% for large inputs.