Avoid memory reads from Vec<Prime> during SIQS
The Prime type is a huge structure (more than 64 bytes).
Pack it into 12 bytes for the critical loop of the sieve.
When the factor base reaches 100k elements, it will fit in about 1MB
instead of overflowing CPU caches.

The speed gain can reach about 5% depending on the CPU cache size
and the size of input.