Use more realistic matrices in benchmatrix
In the actual program, any factor appearing only once among relations
is eliminated since that relation cannot possibly be used in the matrix
kernel.

This also has the advantage of increasing the rank of B^T B which is
crucial for the success of Block Lanczos algorithm.

Enforcing this invariant in the test program allows to successfully
run block Lanczos with smaller block sizes.