Eagerly combine relations
Our cycle finding algorithm is not generic and relies on the
fact that all cycles statistically form a single connected component
(due to the distribution bias toward smaller primes).

This is equivalent to maintaining a spanning tree and converting
any edge connecting 2 tree nodes immediately to a cycle.
To compensate the suboptimality due to greediness, minor updates
are performed when a node is reached by a shorter path.