Fully vectorized SIQS polynomial initialization
Replace the (optimized but) costly reduction modulo p
by a few subtractions: to fully benefit from SIMD and vectorization,
we restore the old array of offsets modulo p and use the array
of u32 primes from the factor base structure.

Since the roots are computed from values already in interval [0,p)
the final result can be obtained by a bounded amount of masked
subtractions by p.

The resulting code can be efficiently expressed by the compiler
as a sequence of SSE2 or AVX2 instructions.

This reduces the cost of SIQS polynomial roots by half.
For very large inputs (> 300 bits) this cost could be about 10%
of total time.