Implement 2-stage Pollard P-1 for small integers
We are interested in obtaining statistically fast factorization
of ~50 bit semiprimes. A study on random primes show that the
2-stage Pollard P-1 method can be very efficient because in virtually
all cases (except strong primes), P-1 has a large prime factor
and the 2nd largest factor is less than 1000.

This is 10x faster than the current implementation of SQUFOF (used as a
fallback) and makes the cost of factoring many double large primes
negligible.