Implement 8k+1 optimization in a consistent way
This optimization is implemented in YAFU and uses the same
idea as Silverman's MPQS article. In Yamaquasi this was used in MPQS
but not in classical QS nor in SIQS.

When n=8k+1, we get statistically smoother numbers by restricting
the sieve to X²-n where X is an odd number. This corresponds to
polynomials Ax²+Bx+C such that B²-4AC = n.

In classical QS the polynomial is ((2x+1)²-n)/4 corresponding to A=1.

The multiplier selection must be modified to generate bias towards
8k+1 values. This is done by recomputing its score taking this
optimization into account. Our calculation gives a score of 3 instead of
YAFU constant 2.625.

The MPQS code already uses this technique but will be refactored
in another commit.

The main effect of this change is a significant increase in relation
yield: in classical QS, a typical speedup is 10%-15%, in SIQS
the speedup can reach 30%.

For example factoring RSA-100 using SIQS will use the very effective
multiplier 139 and complete in about 2 hours instead of 3 hours
on a Ryzen 5500U core.