[siqs] Make each thread handle separate values of A
Previously each value of A was considered as a list of polynomials
to be sieved in parallel by a thread pool. However the number of
polynomials per A value can be small, and the time needed to precompute
data for a given A (on a single thread) can hurt paralellism
significantly (up to 10-20% inefficiency).

This change will also enable further optimizations (memory buffer reuse
and Gray code iteration on polynomials).