ECM implementation revamp
Several changes are included.

The top level program gains a new mode "ecm". This mode does only
ECM, which is not very useful for large numbers but can be used
to test the implementation. However, since curve iteration is
deterministic the recursion to smaller factors is not very effective
because it goes through the same curve.

Implement batch modular inversion for ECM stage 2 and reduce frequency
of intermediate GCD computations. The performance gain is about 30%-40%.

Fix a bug in Montgomery arithmetic.

Use a 1-parameter family of Edwards curve with (2,4)-torsion over Q
as a complement to the EECM family of "good curves". This is slightly
more effective than the previous "random" family which only had a Z/4
torsion subgroup over Q and still provides a nearly infinite supply
of curves.

Modify stage 1 so that the power of 2 and 3 is larger than B1 to take
into account the exponent bias of Edwards curves.

Modify analysis Python script to reflect the changes.

Adjust ECM preliminary step (in auto mode) to take about 1% of
the time the quadratic sieve would take (to ensure overhead remains
small).