Implement Schonhage-Strassen algorithm for convolution of Z/nZ polynomials
The modulus for coefficients is 2^N+1 and the 4N-th root of unity
(2^3n/4-2^n/4) is used for the FFT/NTT.

A FFT of size 2n can be used to compute a full product of degree
n-1 polynomials, or a middle product of degree n-1 and 2n-2
polynomials.

The FFT is combined with Karatsuba multiplication in the frequency
domain to obtain a O(n^1.29) complexity for sizes up to 1 million.