[ecm] Rebalance stage 2 blocks
API is modified to use a target B2 value as argument.
Instead of setting B2=D² which causes a very unbalanced
polynomial evaluation (degree 5φ vs degree φ/2), select
B2 = k d1 such that k is close to a small multiple of φ/2.

The number of "square blocks" is between 1 and 4 as advised
in "20 Years of ECM".

For similar values of B2, balanced blocks can be 10-20% faster.