[ecm] Review parameters and logic for ECM-only mode
The ECM-only mode was using very large number of curves
with rather small parameters. The traditional advice is
to only use the appropriate number of curves for a target
factor size.

By ramping up parameters for increasing factor size, the total
runtime is guaranteed to have a small excess compared to optimal
runtime for the (unknown) factor size.