[pp1] Demo implementation of P+1 algorithm
Using existing code for P-1 and ECM algorithms it is easy
to implement rather efficient code for the P+1 algorithm.

It uses binary Lucas chains for stage 1 and the FFT continuation
through remainder trees for stage 2 (instead of chirp-z transform).

Except in cases where the factor to be found would be suitable
for P+1, the performance is not good enough to be useful compared
to ECM with similar smoothness bounds.