[ecm] Fix runaway complexity in prime sieve
The ECM function prepares exponents for stage 1 using a list
of primes smaller than B1. It was using a costly call to do
that (with an incorrect argument) possibly using nearly 1GB of memory
and >10s of CPU time for B1>=50M.

Use the same implementation as P-1 so that the memory usage
remains closer to B1 bits and CPU cost less than 1 second.

The problem can be visible when invoking ECM multiple times
in parallel from the Python API: the prime base is not shared
across calls, making memory I/O blow up.