[ecm] Implement curve selection from the Suyama-11 family
The Suyama-11 family is an infinite family containing the very effective
Suyama curve σ=11. Its properties are described in article "Finding
ECM-friendly curves through a study of Galois properties"
by Barbulescu-Bos-Bouvier-Kleinjung-Montgomery.

It has a Z/12Z torsion subgroup modulo all primes and can be described
by a genus 1 modular curve, so an infinity of such curves can be
generated by elliptic curve arithmetic (with a small overhead).

This introduces a new class of problems where the rational generator
would have a small order modulo the prime factors. In that case
we fall back to more generic Edwards curve hoping that they expose
the factors.

The theoretical study shows that it has the same smoothness properties
as p/30 (instead of p/16 for the Z/2 x Z/4 Edwards curves) making
them almost as effective as the EECM good curves (about 10% computation
cost improvement).

The ECM parameter optimization script is updated with this new value.