[ecm] Correct/expand study scripts
The addition chain simulations are refactored and simplified
(for large blocks, we don't need the finely optimized chains
that make a difference for small exponents).

The ECM efficiency study now distinguishes curve torsion statistics
depending on the residue classes mod 12. Some of these residue
classes are correlated with higher 2-torsion or 3-torsion.