Provide an implementation of Pollard's rho algorithm
Pollard's rho algorithm is space efficient and does not require
quotients or square roots like SQUFOF.

Compared to the existing "small P-1" which was using suboptimal
parameters, it is both faster and more efficient with success rates
higher than 90-95% (instead of 85-95% or even 70% for 56-bit inputs).

It will make it possible to increase the quantity of double-large-primes
in quadratic sieve without too much performance impact.