Remove multiprecision P-1 and ECM from 64-bit algorithm autoselection
Algorithms written for multiprecision integers have a large performance
gap with algorithms written for u64 inputs.

It is faster to simply use Pollard's rho by default for 64-bit inputs
and not attempt generic P-1 and ECM at all. In the random integer
benchmark, this reduces execution time by 20-50% for small integers.

It is still possible to run them manually.