[siqs] Fix edge case where enough (but not many) relations are found
The SIQS implementation requires more than 64 relations in excess
of the factor base size in order to make block Lanczos succeed.

Small inputs may not find that amount of relations: but Gauss reduction
would work as long as there is a single excess relation. Tolerate
that situation to avoid spurious failures.