[mpqs] Major overhaul of implementation
The MPQS implementation was written early and can be improved
by imitating the SIQS implementation, since they should share
most of the logic except polynomial selection.

For small integers, incorrect logic was preventing the code
from selecting optimal A (it could not be less than 200 and not be part
of the factor base).

The following changes are made:
- properly honor verbosity flag
- representation of polynomial as ax²+bx+c instead of (ax+b)²-n
- stop using large primes for small integers
- larger intervals for large inputs
- common state object for the whole sieve process
- avoid repeated calls to prepare_prime during sieve
- use u128 to represent parameter D = sqrt(A)
- reduce factor base when using double large primes (as done in SIQS)
- use root location hint for sieve threshold
- allow for more hashtable overflows during sieve (due to larger intervals)

This allows to recover an appropriate algorithm complexity.

Benchmark results for random semiprimes (Ryzen 5500U, Zen2 4GHz)

          Before    After
40 bits   9.30ms    0.45ms
48 bits   4.40ms    0.50ms
56 bits   2.94ms    0.59ms
64 bits   2.42ms    0.70ms
72 bits   2.46ms    0.88ms
80 bits   2.71ms    1.21ms
88 bits   3.30ms    1.81ms
96 bits   4.45ms    3.08ms
128 bits  25.8ms    22.5ms
160 bits   245ms     210ms