New specialized 128-bit ECM implementation
It is modelled on the existing ECM implementation with
the following differences:
- it uses its own modular arithmetic for 128-bit moduli
- it only uses twisted Edwards curves
- it cannot use large exponent blocks or FFT continuation

The parameters are tailored to be optimal for products
of balanced primes, and cover the range from 64 to 128 bits.

The ECM128 is faster for integers from 65 to 80 bits.
It may be expanded to cover the 60-64 bit range where
it possibly beats the Pollard's rho implementation.

As a consequence this change makes factoring 66 bit integers
faster (using ECM128) than 64-bit integers (relying on Rho algorithm).