Estimate progress percentage when logging about quadratic sieve
Using the current number of relations it is possible to predict
quite accurately when quadratic sieve will complete. This is
because relations appear at a linear rate, cycles of length 2
(coming from p-relations) are a quadratic function of time,
and relations from double large primes appear at rate O(t^α)
where α is between 3 and 4.

These approximations are valid whenever the number of relations
is much smaller than the space of large primes (which is always true).