[siqs] Implement Gray code iteration when switching polynomials
The self-initializing MPQS works by selecting A to be a product
of small primes and B to be one of the 2^k modular square roots
of n mod A.

The traditional method to iterate over these roots is to walk
through them using a binary Gray code. It was not implemented
because it is less compatible with parallel processing.

Now that parallelization is done on A values, each thread can
iterate efficiently on polynomials.

Since the Rust+LLVM toolchain already has efficient vectorization,
remove clumsy handwritten vectorization. Using native min/max
operations, 3 AVX2 instructions are enough to process a block
of 8 roots.

This reduces the cost of polynomial preparation ~3-fold.