Use a mixed dense-sparse representation during block Lanczos
A similar strategy is used in msieve: it is effective because
quadratic matrices are very unbalanced with nearly half of their
coefficients in first few rows.

The 'benchmatrix' benchmark is too different from actual
matrices to properly simulate the performance improvement.

The speed improvement for the linear algebra part is about 2x-2.5x.
It is especially relevant in multithreaded mode where the
single-threaded linear algebra can be up to 5%-10% of wall time.