Recycle allocated memory of Sieve objects
For large inputs, the hash tables allocated during sieve
use more than twice the interval size (in bytes).
In multithreaded mode, this can quickly fill up caches
especially if a new table is allocated and zeroed.

We can reuse tables to avoid this step.

Depending on cache size, number of threads, and pressure on
the memory controller, the observed cost of zeroing the
SieveTable objects could range from 1% to 7% of CPU time.