[siqs] Fix calibration script
The calibration script is also used to measure performance
during development. It was not recycling Sieve objects,
leading to incorrect and unstable performance estimates.

Stop testing multiple large prime thresholds: it is already clear
that very large values have no benefit for convergence speed,
and the main purpose of calibration is to select the correct
factor base and interval sizes.

Instead it will now test a larger range of interval sizes.