Remove generic closure from sieve core function
For historical reasons a generic closure is stored in the Sieve
object to handle QS/MPQS/SIQS using a single copy of the code.
This is artificially generic and it is just as easy to pass
a slice of roots as arguments.

This allows to focus the implementation on SIQS, even if the other
(educational) implementations become a bit less efficient.

A few instructions are saved in the process but this is not
the primary purpose of the change.