Calibration scripts for quadratic sieve
It is possible to optimize the factor base sizes
using the asymptotic smoothness probability given
by the Dickman rho function and its variants, but
this is more complex if we want to take into account
the (double) large prime variations and the multiplier.

Real life benchmarks are easier to implement and can also measure
the dependence on cache size.