[qs] Simpler and faster classical quadratic sieve
The classical QS uses a single polynomial (r+x)²-n
and the implementation considers successive intervals
corresponding to polynomials (r+x+offset)²-n.

Instead of recomputing the roots modulo the prime base each time
it is simpler to add the width of the interval each time (it can
be vectorized like the roots computation in SIQS).

The block size was too large, making memory locality very poor.
Also, a significant part of the sieve was not parallel:
we can have about 180%/190% CPU usage instead of 135% in 2-thread mode.

The factor base argument was not properly handled when using the double
large prime variation (it was silently halved).

The quadratic sieve becomes usually 2x faster (when using 2 threads).